condominioaldia/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.3
django-allauth==0.38.0
django-countries==5.3.2
django-discover-runner==1.0
django-filter==2.0.0
django-import-export==1.1.0
django-libs==2.0.2
django-money==0.14.3
django-rest-auth==0.9.3
django-role-permissions==2.2.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
Markdown==3.0.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.10
Pillow==5.3.0
py-moneyed==0.7.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
